Refactor: 타인 기록은 바로 열리도록 처리
